refactor(config): isolate runtime load balancing policy interpretation
